doc-summarizer-qa/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
python-multipart==0.0.6
PyPDF2==3.0.1
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1
psycopg2-binary>=2.9.11
sqlalchemy>=2.0.36
python-dotenv>=1.0.0
